파일정리&라즈베리파이 프로그램

+++ source/rpi.py
@@ -5,37 +5,60 @@ import requests
 from picamera import PiCamera
 from time import sleep
 import os
+import cv2 as cv
 
+face_classifier = cv.CascadeClassifier('haarcascade_frontalface_default.xml')
+
+def FaceExtract(img):
+    gray = cv.cvtColor(img, cv.COLOR_BGR2GRAY)
+    faces = face_classifier.detectMultiScale(gray,1.3,5)
+    if faces is ():
+        return None
+    
+    for (x, y, w, h)in faces:
+        cropped_face = img[y:y+h, x:x+w]
+        
+    return cropped_face
+    
 def CapData():
     count = 0
     global requestUrl
     
+    cam = PiCamera()
     cam.start_preview()
+    sleep(1)
+    
     while True:
         filename = str(count)+'.jpg'
-        cam = PiCamera()
+        cam.capture('temp.jpg')
+        frame = cv.imread('temp.jpg')
+        if FaceExtract(frame) is not None:
+            count += 1
+            face = cv.resize(FaceExtract(frame), (200, 200))
+            face = cv.cvtColor(face, cv.COLOR_BGR2GRAY)
+            cv.imwrite(filename, face)
+        else:
+            print(str(count)+"Face Not Fount")
+            os.remove('temp.jpg')
+            continue
+            
+        if cv.waitKey(1) == 13 or count == 100:
+            break
         
-        sleep(1)
-        cam.capture(filename)
         files = {'file': open(filename, 'rb')}
         response = requests.post(requestUrl+'/upload', files=files)
+        os.remove('temp.jpg')
         os.remove(filename)
         
         if response.status_code == 200:
-            print(count+"file Send Success")
+            print(str(count)+"file Send Success")
         
         if count == 100:
             break
         
     cam.stop_preview()
-    
-
+    cam
 
 if __name__ == '__main__':
     requestUrl = "http://192.168.0.43:5000"  
-    while True:
-        resp = requests.post(requestUrl)
-        if resp == "nodata":
-            CapData()
-            
-        sleep(1)
+    CapData()
